# pup_patrol_backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pup_patrol",
        "version": "1"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "posts 테이블",
                "description": "posts 테이블에 등록된 모든 데이터를 data라는 키를 가진 배열로 반환",
                "operationId": "e9eaf7df702a94f4443677bd1a36cfe6",
                "responses": {
                    "200": {
                        "description": "success"
                    }
                }
            },
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "posts 테이블에 데이터 추가",
                "description": "posts 테이블에 데이터를 추가함",
                "operationId": "4a657e51a0066c64781851f2fd2651e2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "postTitle": {
                                        "description": "게시글 제목",
                                        "type": "string",
                                        "example": "title"
                                    },
                                    "postContent": {
                                        "description": "게시글 내용",
                                        "type": "string",
                                        "example": " [\\'ddd\\',\\'(IMAGE)1725807499_pome.jpg\\',\\'ddd\\'] (배열 형태의 텍스트여야 함, 배열의 요소 중 (IMAGE) 가 붙은 건 이미지 이므로 앞의 (IMAGE)를 떼고 남은 파일 이름을 api/imageDownload 에 요청을 보내서 이미지 링크로 받아와야 함)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "게시글이 성공적으로 추가됨"
                    }
                }
            }
        },
        "/api/posts/{post}": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "파라미터로 넘어온 id를 가지는 특정 post 데이터 조회",
                "description": "파라미터로 넘어온 id를 가지는 특정 post 데이터를 data라는 키를 가진 객체로 반환",
                "operationId": "d3910995cb542d02254c9fc6133cc10c",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "post의 id",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "하나의 post 데이터"
                    }
                }
            },
            "put": {
                "tags": [
                    "post"
                ],
                "summary": "post데이터 수정",
                "description": "로그인 한 정보와 글 작성자가 같다면 post데이터를 수정함 ",
                "operationId": "4897f3aff4c2ef880950782e3669baf4",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "post의 아이디",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "postTitle": {
                                        "description": "게시글 제목",
                                        "type": "string",
                                        "example": "title"
                                    },
                                    "postContent": {
                                        "description": "게시글 내용",
                                        "type": "string",
                                        "example": "[\\'ddd\\',\\'(IMAGE)1725807499_pome.jpg\\',\\'ddd\\'] (배열 형태의 텍스트여야 함, 배열의 요소 중 (IMAGE) 가 붙은 건 이미지 이므로 앞의 (IMAGE)를 떼고 남은 파일 이름을 api/imageDownload 에 요청을 보내서 이미지 링크로 받아와야 함)"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "수정이 된 경우"
                    },
                    "401": {
                        "description": "로그인된 정보와 글 작성자가 일치하지 않을 경우"
                    }
                }
            },
            "delete": {
                "tags": [
                    "post"
                ],
                "summary": "특정 post 데이터 삭제",
                "description": "로그인한 정보와 글 작성자가 같다면 post를 삭제함",
                "operationId": "995630fb8f7510fb391745db0c073d23",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "description": "post의 id",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "post 삭제에 성공한 경우"
                    },
                    "401": {
                        "description": "로그인된 정보와 글 작성자가 일치하지 않을 경우"
                    }
                }
            }
        },
...
        "/api/questions/{question}": {
            "get": {
                "tags": [
                    "question"
                ],
                "summary": "특정 question 데이터 조회",
                "description": "파라미터로 받은 id를 가지는 question 객체를 data라는 키를 가진 객체로 반환",
                "operationId": "8bc2cdefaa6d0469ec03ceb8352077a5",
                "parameters": [
                    {
                        "name": "question",
                        "in": "path",
                        "description": "question의 id",
                        "required": true,
                        "schema": {
                            "type": "json"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "한 개의 question 데이터"
                    }
                }
            }
        }
    },
...
}